+++ labs/lab4.sh
@@ -37,6 +37,9 @@ check() {
 
 # lock lab
 lock() {
+    if [ -n "$DEBUG" ]; then
+        return 0
+    fi
     files=(/tmp/lab*)
     if [ -e "${files[0]}" ]; then
         base_name=$(basename "${files[0]}")
@@ -49,6 +52,9 @@ lock() {
 
 # unlock lab
 unlock() {
+    if [ -n "$DEBUG" ]; then
+        return 0
+    fi
     rm -f "/tmp/$ME" || true
 }
